feat: move into POST /checkout from POST /orders
