+++ src/modules/checkout/route.ts
@@ -1,8 +1,9 @@
 import { createRoute, OpenAPIHono } from "@hono/zod-openapi";
 import { prisma } from "@/lib/prisma";
-import { ErrorResponseSchema } from "@/modules/common/schema";
-import { CheckoutResponseSchema } from "@/modules/checkout/schema";
+import { CreateCheckoutBodySchema, GetCheckoutResponseSchema } from "@/modules/checkout/schema";
 import { Env } from "@/index";
+import { OrderSchema } from "@/modules/order/schema";
+import { ErrorResponseSchema } from "../common/schema";
 
 export const checkoutRoute = new OpenAPIHono<Env>();
 const tags = ["Checkout"];
@@ -11,15 +12,14 @@ const tags = ["Checkout"];
 checkoutRoute.openapi(
   createRoute({
     tags,
-    summary:
-      "Get checkout page data (profile, cart, address, shipping-methods, payment-methods)",
+    summary: "Get checkout page data (profile, cart, address, shipping-methods, payment-methods)",
     method: "get",
     path: "/checkout",
     responses: {
       401: { description: "Unauthorized" },
       200: {
         description: "Successfully fetched all checkout data",
-        content: { "application/json": { schema: CheckoutResponseSchema } },
+        content: { "application/json": { schema: GetCheckoutResponseSchema } },
       },
       500: { description: "Server error" },
     },
@@ -74,6 +74,102 @@ checkoutRoute.openapi(
         shippingMethods,
         paymentMethods,
       });
+    } catch (error) {
+      console.error("[GET /checkout] error:", error);
+      return c.json({ message: "Failed to load checkout data", error }, 500);
+    }
+  }
+);
+
+// POST /checkout
+checkoutRoute.openapi(
+  createRoute({
+    tags,
+    summary:
+      "Checkout to validate customer info, shipping address, shipping method, payment method; before create new order",
+    method: "post",
+    path: "/checkout",
+    request: { body: { content: { "application/json": { schema: CreateCheckoutBodySchema } } } },
+    responses: {
+      201: {
+        description: "Successfully checked out and created new order",
+        content: { "application/json": { schema: OrderSchema } },
+      },
+      401: { description: "Unauthorized" },
+      404: {
+        content: { "application/json": { schema: ErrorResponseSchema } },
+        description: "Cart not found",
+      },
+      500: { description: "Server error" },
+    },
+  }),
+  async (c) => {
+    const user = c.get("user");
+    if (!user) return c.text("Unauthorized", 401);
+    const body = c.req.valid("json");
+
+    try {
+      const [cart, address, shippingMethod, paymentMethod] = await Promise.all([
+        prisma.cart.findFirst({
+          where: { userId: user.id },
+          include: { items: { include: { product: { include: { images: true } } } } },
+        }),
+        prisma.address.findFirst({
+          where: { id: body.addressId, userId: user.id },
+        }),
+        prisma.shippingMethod.findUnique({
+          where: { slug: body.shippingMethodSlug },
+        }),
+        prisma.paymentMethod.findUnique({
+          where: { slug: body.paymentMethodSlug },
+        }),
+      ]);
+
+      if (!cart || cart.items.length === 0) return c.json({ message: "Cart not found or empty" }, 404);
+      if (!address) return c.json({ message: "Address not found" }, 404);
+      if (!shippingMethod) return c.json({ message: "Shipping method not found" }, 404);
+      if (!paymentMethod) return c.json({ message: "Payment method not found" }, 404);
+
+      // TODO: Validate data completion of user and address
+
+      // Finally create the order
+      const orderNumber = `ORD-${Date.now()}-${Math.random().toString(36).substring(2, 8).toUpperCase()}`;
+      const shippingCost = shippingMethod.price;
+      const totalAmount = cart.totalPrice + shippingCost;
+
+      const order = await prisma.order.create({
+        data: {
+          status: "PENDING",
+          orderNumber,
+          userId: user.id,
+          shippingAddressId: address.id, // TODO: will copy the address
+          notes: address.notes || "",
+          courier: null,
+          trackingNumber: null,
+          paymentMethodId: paymentMethod.id,
+          orderItems: {
+            createMany: {
+              data: cart.items.map((item) => ({
+                productId: item.productId,
+                quantity: item.quantity,
+                price: item.product.price,
+                total: item.quantity * item.product.price,
+              })),
+            },
+          },
+          subTotal: cart.totalPrice,
+          shippingCost,
+          discount: 0,
+          totalAmount,
+        },
+        include: {
+          orderItems: { include: { product: { include: { images: true } } } },
+          shippingAddress: true,
+          paymentMethod: true,
+        },
+      });
+
+      return c.json(order, 201);
     } catch (error) {
       console.error("[GET /checkout] error:", error);
       return c.json({ message: "Failed to load checkout data", error }, 500);
