+++ src/modules/order/schema.ts
@@ -1,8 +1,5 @@
 import { z } from "zod";
-import {
-  OrderSchema as BaseOrderSchema,
-  OrderItemSchema as BaseOrderItemSchema,
-} from "@/generated/zod";
+import { OrderSchema as BaseOrderSchema, OrderItemSchema as BaseOrderItemSchema } from "@/generated/zod";
 import { PaymentMethodSchema } from "@/modules/payment-method/schema";
 import { AddressSchema } from "@/modules/address/schema";
 import { ProductSchema } from "@/modules/product/schema";
@@ -21,27 +18,12 @@ export const OrderSchema = BaseOrderSchema.extend({
   orderItems: z.array(OrderItemSchema),
 });
 
-// Create Order Schema (for POST request)
-export const CreateNewOrderSchema = z.object({
-  addressId: z.string(),
-  shippingMethodSlug: z.string(),
-  paymentMethodSlug: z.string(),
-  // User Profile and Latest Cart from database
-});
-
 // Order List Schema
 export const OrderListSchema = z.array(OrderSchema);
 
 // Update Order Status Schema
 export const UpdateOrderStatusSchema = z.object({
-  status: z.enum([
-    "pending",
-    "confirmed",
-    "processing",
-    "shipped",
-    "delivered",
-    "cancelled",
-  ]),
+  status: z.enum(["pending", "confirmed", "processing", "shipped", "delivered", "cancelled"]),
 });
 
 // Param Schema
